adding ktx dependency and latest navigation component
