+++ app/src/main/java/com/andor/navigate/notepad/expanded/ExpandedNoteFragment.kt
@@ -7,7 +7,7 @@ import android.view.*
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import androidx.navigation.Navigation
+import androidx.navigation.fragment.NavHostFragment.findNavController
 import com.andor.navigate.notepad.R
 import com.andor.navigate.notepad.core.NoteViewModel
 import com.andor.navigate.notepad.listing.NotesActivity
@@ -49,7 +49,7 @@ class ExpandedNoteFragment : Fragment() {
         if (item.itemId == R.id.edit) {
             val action =
                 ExpandedNoteFragmentDirections.actionExpandedNoteFragmentToUpdateNoteFragment(true)
-            Navigation.findNavController(view!!).navigate(action)
+            findNavController(this).navigate(action)
         }
         return super.onOptionsItemSelected(item)
     }
