--- app/src/main/java/com/andor/navigate/notepad/listing/fragment/BottomSheetNavFragment.kt
@@ -1,22 +1,0 @@
-package com.andor.navigate.demonavigation.fragment
-
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.andor.navigate.notepad.R
-
-class BottomSheetNavFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_bottom_nav, container, false)
-    }
-
-
-}
