+++ app/src/main/java/com/andor/navigate/notepad/listing/fragment/NoteListingFragment.kt
@@ -8,13 +8,15 @@ import android.view.inputmethod.InputMethodManager
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import androidx.navigation.Navigation
+import androidx.navigation.fragment.NavHostFragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.andor.navigate.notepad.R
-import com.andor.navigate.notepad.core.*
+import com.andor.navigate.notepad.core.AppState
+import com.andor.navigate.notepad.core.ListingType
+import com.andor.navigate.notepad.core.NoteViewModel
 import com.andor.navigate.notepad.listing.NotesActivity
 import com.andor.navigate.notepad.listing.adapter.ListItemEvent
 import com.andor.navigate.notepad.listing.adapter.ListingAdapter
@@ -24,7 +26,6 @@ import kotlinx.android.synthetic.main.fragment_note_listing.*
 
 class NoteListingFragment : Fragment() {
     private lateinit var oldAppState: AppState
-    private val bottomSheetMenuFragment = BottomSheetMenuFragment()
     private var longPressActionMode: androidx.appcompat.view.ActionMode? = null
     private var isLongPressed: Boolean = false
     private val selectedNotes: HashSet<NoteModel> = HashSet()
@@ -88,14 +89,6 @@ class NoteListingFragment : Fragment() {
         }
     }
 
-    private fun sendAddNoteBottomSheetCommand(menuType: BottomMenuType) {
-        if (!bottomSheetMenuFragment.isAdded) {
-            viewModel.openBottomMenu(menuType)
-        } else {
-            bottomSheetMenuFragment.dismiss()
-        }
-    }
-
     private fun hideKeyBoard() {
         val imm = activity!!.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         imm.hideSoftInputFromWindow(view!!.windowToken, 0)
@@ -105,33 +98,9 @@ class NoteListingFragment : Fragment() {
         viewModel.appStateRelay.observe(this, Observer { appState ->
             appState?.let { notNullAppState ->
                 updateRecyclerView(notNullAppState)
-                handleBottomSheetEvent(notNullAppState)
-
                 oldAppState = appState
             }
         })
-    }
-
-    private fun handleBottomSheetEvent(appState: AppState) {
-        if (::oldAppState.isInitialized && oldAppState.bottomMenuEvent != appState.bottomMenuEvent) {
-            when (appState.bottomMenuEvent) {
-                is BottomMenuEvent.AddNote -> {
-                    bottomSheetMenuFragment.dismiss()
-                    Navigation.findNavController(view!!).navigate(R.id.action_noteListingFragment_to_updateNoteFragment)
-                }
-                is BottomMenuEvent.Close -> bottomSheetMenuFragment.dismiss()
-                is BottomMenuEvent.Open -> {
-                    if (!bottomSheetMenuFragment.isAdded) {
-                        bottomSheetMenuFragment.show(
-                            activity!!.supportFragmentManager,
-                            "Bottom_Sheet"
-                        )
-                    } else {
-                        bottomSheetMenuFragment.dismiss()
-                    }
-                }
-            }
-        }
     }
 
     private fun updateRecyclerView(
@@ -160,7 +129,7 @@ class NoteListingFragment : Fragment() {
                     viewModel.updateSelectedNotes(it.noteModel)
                     val action =
                         NoteListingFragmentDirections.actionNoteListingFragmentToExpandedNoteFragment()
-                    Navigation.findNavController(view!!).navigate(action)
+                    findNavController(this).navigate(action)
                 } else {
                     if (selectedNotes.contains(it.noteModel)) {
                         selectedNotes.remove(it.noteModel)
@@ -205,10 +174,10 @@ class NoteListingFragment : Fragment() {
             longPressActionMode?.finish()
         }
         if (item.itemId == R.id.action_setting) {
-            sendAddNoteBottomSheetCommand(BottomMenuType.Setting)
+            findNavController(this).navigate(R.id.action_noteListingFragment_to_settingFragment)
         }
         if (item.itemId == R.id.action_add) {
-            sendAddNoteBottomSheetCommand(BottomMenuType.AddNote)
+            findNavController(this).navigate(R.id.action_noteListingFragment_to_addNewNoteFragment)
         }
         return super.onOptionsItemSelected(item)
     }
