+++ app/src/main/res/layout/fragment_add_new_note.xml
@@ -2,9 +2,11 @@
 
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/addFragmentContainer"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    tools:context=".listing.fragment.AddNewNoteFragment"
     android:orientation="vertical">
 
     <TextView
