--- app/src/main/res/layout/fragment_bottom_sheet_menu.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/bottomMenuContainer"
-    tools:context=".listing.fragment.BottomSheetMenuFragment">
-
-</FrameLayout>
