+++ app/src/main/res/navigation/nav_graph.xml
@@ -24,6 +24,12 @@
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
             app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+        <action
+            android:id="@+id/action_noteListingFragment_to_settingFragment"
+            app:destination="@id/settingFragment" />
+        <action
+            android:id="@+id/action_noteListingFragment_to_addNewNoteFragment"
+            app:destination="@id/addNewNoteFragment" />
     </fragment>
     <fragment
         android:id="@+id/expandedNoteFragment"
@@ -51,6 +57,26 @@
             android:name="editMode"
             android:defaultValue="false"
             app:argType="boolean" />
-
     </fragment>
+
+    <dialog
+        android:id="@+id/settingFragment"
+        android:name="com.andor.navigate.notepad.listing.fragment.SettingFragment"
+        tools:layout="@layout/fragment_setting">
+        <action
+            android:id="@+id/action_settingFragment_to_confirmationFragment"
+            app:destination="@id/confirmationFragment" />
+    </dialog>
+    <dialog
+        android:id="@+id/addNewNoteFragment"
+        android:name="com.andor.navigate.notepad.listing.fragment.AddNewNoteFragment"
+        tools:layout="@layout/fragment_add_new_note" >
+        <action
+            android:id="@+id/action_addNewNoteFragment_to_updateNoteFragment"
+            app:destination="@id/updateNoteFragment" />
+    </dialog>
+    <dialog
+        android:id="@+id/confirmationFragment"
+        android:name="com.andor.navigate.notepad.listing.fragment.ConfirmationFragment"
+        tools:layout="@layout/fragment_confirmation" />
 </navigation>
