--- app/src/main/res/navigation/nav_graph_bottom_menu.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_graph_bottom_menu"
-    app:startDestination="@id/navMenuFragment">
-
-    <fragment
-        android:id="@+id/addNewNoteFragment"
-        android:name="com.andor.navigate.notepad.listing.fragment.AddNewNoteFragment"
-        android:label="fragment_add_new_note"
-        tools:layout="@layout/fragment_add_new_note" />
-    <fragment
-        android:id="@+id/navMenuFragment"
-        android:name="com.andor.navigate.notepad.listing.fragment.NavMenuFragment"
-        android:label="fragment_nav_menu"
-        tools:layout="@layout/fragment_nav_menu" >
-        <action
-            android:id="@+id/action_navMenuFragment_to_settingFragment"
-            app:destination="@id/settingFragment2"
-            app:launchSingleTop="false"
-            app:popUpToInclusive="true" />
-        <action
-            android:id="@+id/action_navMenuFragment_to_addNewNoteFragment"
-            app:destination="@id/addNewNoteFragment"
-            app:launchSingleTop="false"
-            app:popUpToInclusive="true" />
-    </fragment>
-    <fragment
-        android:id="@+id/settingFragment2"
-        android:name="com.andor.navigate.notepad.listing.fragment.SettingFragment"
-        android:label="fragment_setting"
-        tools:layout="@layout/fragment_setting" />
-</navigation>
