+++ app/src/main/res/values/strings.xml
@@ -24,4 +24,7 @@
     <string name="logout">Logout</string>
     <string name="list_view_type">List View Type</string>
     <string name="add_note">Add Note</string>
+    <string name="yes">Yes</string>
+    <string name="no">No</string>
+    <string name="signout">SignOut?</string>
 </resources>
