+++ build.gradle
@@ -2,6 +2,13 @@
 
 buildscript {
     ext.kotlin_version = '1.3.31'
+    ext.roomVersion = '2.1.0'
+    ext.archLifecycleVersion = '2.2.0-alpha01'
+    ext.androidxArchVersion = '2.0.0'
+    ext.coroutines = '1.2.0'
+    ext.nav_version = "2.1.0-alpha05"
+
+
     repositories {
         google()
         jcenter()
@@ -10,19 +17,13 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:3.4.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.1.0-alpha04"
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
 
         //fire-base dependency
         classpath 'com.google.gms:google-services:4.2.0'
     }
-}
-ext {
-    roomVersion = '2.1.0-rc01'
-    archLifecycleVersion = '2.2.0-alpha01'
-    androidxArchVersion = '2.0.0'
-    coroutines = '1.2.0'
 }
 
 allprojects {
